+++ sdk/identity/Azure.Identity/src/AuthenticationRequiredException.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Runtime.Serialization;
 using Azure.Core;
 
 namespace Azure.Identity
@@ -9,6 +10,7 @@ namespace Azure.Identity
     /// <summary>
     /// An exception indicating that interactive authentication is required.
     /// </summary>
+    [Serializable]
     public class AuthenticationRequiredException : CredentialUnavailableException
     {
         /// <summary>
@@ -31,6 +33,16 @@ namespace Azure.Identity
             : base(message, innerException)
         {
             TokenRequestContext = context;
+        }
+
+        /// <summary>
+        /// A constructor used for serialization.
+        /// </summary>
+        /// <param name="info">The <see cref="SerializationInfo"/>.</param>
+        /// <param name="context">The <see cref="StreamingContext"/>.</param>
+        /// <returns></returns>
+        protected AuthenticationRequiredException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
         }
 
         /// <summary>